refactoring: move getPackageFile to package file module
